refactor(cli): split core.py into focused command modules
Phase 3 (CLI code health) PR-B. Behavior-preserving split of the
1166-line cli/commands/core.py into focused modules; core.py becomes an
18-line aggregator that re-exports the Typer app.

New modules under cli/commands/:
- _base.py        app instance, console, version command, _main callback
- _helpers.py     shared _write_output
- compile_cmd.py  _run_compile, compile, batch
- transform.py    fix, compare, pack
- validation.py   validate
- json_tools.py   json-path, diff

Registration uses the _base.py app-instance pattern (leaf modules import
app from _base and register via @app.command(); core.py imports them for
side effects), so there is no circular import and
'from cli.commands.core import app' keeps working.

Also removes dead code carried in core.py: the never-read
_suppress_log setattrs in batch, the commented-out history integration.
rag.py keeps its own _write_output (different precedence) per the design.

No behavior change: the 9 top-level commands, flags, and output are
preserved. All 58 tests across the 8 cli-importing test files pass
(command-surface invariant + fix/compare/pack characterization +
phase1/2/console-refactor/output-format/optimizer-history/pr-safety),
smoke output identical, ruff v0.1.14 clean.
